turn custom styles into style sheets

+++ viscid/plot/mpl_style.py
@@ -14,11 +14,7 @@ Attributes:
         unpacked into the rc function call.
 
 Style Sheets:
-    The following style sheets are available to matplotlib. Style
-    sheets require matplotlib version >= 1.5.0
-
-    - **viscid-default**: use afmhot
-    - **viscid-steve**: use Steve's label formatter for colorbars
+    See :doc:`mpl_style` for a list of custom Viscid style sheets
 
 Rc Parameters:
     Here are some viscid-specific options that can be put into the
@@ -38,6 +34,9 @@ Rc Parameters:
 
 from __future__ import division, print_function
 from distutils.version import LooseVersion
+from glob import glob
+import os
+import tempfile
 import re
 
 import matplotlib
@@ -69,47 +68,44 @@ for key, default_converter in viscid_mpl_rc_params.items():
     matplotlib.rcParamsOrig[key] = default_converter[0]
     matplotlib.rcParamsDefault[key] = default_converter[0]
 
-VISCID_STYLES = {
-    #######################
-    u"viscid-default": u"""
-image.cmap: viridis
-viscid.symmetric_cmap: RdBu_r
-""",
-    u"viscid-colorblind": u"""
-axes.prop_cycle: cycler('color', ['004358', 'FD7400', '3DA88E', \
-                                  '83522B', '00D4FD', 'E2D893'])
-""",
-    #####################
-    u"viscid-steve": u"""
-viscid.cbarfmt: steve
-"""
-}
 
-# hack the cycler stuff back to the pre-1.5.0 syntax
-if LooseVersion(matplotlib.__version__) < LooseVersion("1.5.0"):
-    for k, v in VISCID_STYLES.items():
-        v = re.sub(r"(.*?)\.prop_cycle\s*:\s*cycler\((['\"])(.*?)\2,\s*"
-                   r"\[\s*(.*)\s*\]\s*\)",
-                   r"\1.\3_cycle: \4", v)
-        v = v.replace("'", "").replace('"', "")
-        VISCID_STYLES[k] = v
-
+def _cycler2prop_cycle(txt):
+    """turn matplotlib 1.5.0 cycler syntax back to old syntax"""
+    txt = re.sub(r"(.*?)\.prop_cycle\s*:\s*cycler\((['\"])(.*?)\2,\s*"
+                 r"\[\s*(.*)\s*\]\s*\)",
+                 r"\1.\3_cycle: \4", txt)
+    txt = txt.replace("'", "").replace('"', "")
+    return txt
 
 def inject_viscid_styles(show_warning=True):
     try:
-        import tempfile
-        import os
-
         from matplotlib import rc_params_from_file, style
 
-        tmpfname = tempfile.mkstemp()[1]
-        for style_name, text in VISCID_STYLES.items():
-            with open(tmpfname, 'w') as f:
-                f.write(text)
-            params = rc_params_from_file(tmpfname, use_default_template=False)
-            style.library[style_name] = params
-        os.unlink(tmpfname)
+        styl_dir = os.path.realpath(os.path.dirname(__file__))
+        styl_dir = os.path.abspath(os.path.join(styl_dir, "styles"))
+        style_sheets = glob(os.path.join(styl_dir, "*.mplstyle"))
+
+        if LooseVersion(matplotlib.__version__) < LooseVersion("1.5.0"):
+            tmpfname = tempfile.mkstemp()[1]
+        else:
+            tmpfname = None
+
+        for styl_fname in style_sheets:
+            styl_name = os.path.splitext(os.path.basename(styl_fname))[0]
+
+            if tmpfname:
+                # hack the cycler stuff back to the pre-1.5.0 syntax
+                with open(styl_fname, 'r') as fin:
+                    with open(tmpfname, 'w') as fout:
+                        fout.write(_cycler2prop_cycle(fin.read()))
+                styl_fname = tmpfname
+            params = rc_params_from_file(styl_fname, use_default_template=False)
+            style.library[styl_name] = params
+
+        if tmpfname:
+            os.unlink(tmpfname)
         style.reload_library()
+
     except ImportError:
         if show_warning:
             logger.debug("Upgrade to matplotlib >= 1.5.0 to use style sheets")
